Added new notebook job icons, fixed a minor bug in openmaterializednotebookAPI and added fixed the schedule Picker API.

+++ extensions/agent/src/mainController.ts
@@ -99,8 +99,9 @@ export class MainController {
 		});
 		vscode.commands.registerCommand('agent.openNotebookEditorFromJsonString', async (filename: string, jsonNotebook: string) => {
 			const tempfilePath = path.join(os.tmpdir(), filename + '.ipynb');
-
-			fs.writeFileSync(tempfilePath, jsonNotebook);
+			if (!fs.existsSync(tempfilePath)) {
+				fs.writeFileSync(tempfilePath, jsonNotebook);
+			}
 			let uri = vscode.Uri.file(tempfilePath);
 			await azdata.nb.showNotebookDocument(uri);
 		});
